refactor: move logging tests into a new file
